comments

+++ Blog/views.py
@@ -1,5 +1,10 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from django.core.paginator import Paginator
+
+from django.contrib.contenttypes.models import ContentType
+from comments.forms import CommentForm
+from comments.models import Comment
+
 from .models import Post
 from .forms import PostForm
 # Create your views here.
@@ -30,15 +35,62 @@ def all_posts(request):
 
     return render(request, "all_posts.html", { 'posts': posts })    
 # post_item
+    # post = Post.objects.get(id=id)
+    # content_type = ContentType.objects.get_for_model(Post)
+    # obj_id = post.id
+    # comments = Comment.objects.filter(content_type = content_type, object_id = obj_id)
+    # instance = get_object_or_404(Post, id=id)
+    # comments = Comment.objects.filter_by_instance(instance)
 def post_item(request, id):
+    instance = get_object_or_404(Post, id=id)
+    initial_data = {
+        "content_type": instance.get_content_type,
+        "object_id": instance.id
+    }
+    form = CommentForm(request.POST or None, initial = initial_data)
+    
+    if form.is_valid():
+        c_type = form.cleaned_data.get("content_type")
+        content_type = ContentType.objects.get(model = c_type)
+        obj_id = form.cleaned_data.get("object_id")
+        content_data = form.cleaned_data.get("content")
+        parent_obj = None
 
-    post = Post.objects.get(id=id)
+        try:
+            parent_id = int(request.POST.get("parent_id"))
+        except:
+            parent_id = None
+        if parent_id:
+            parent_qs = Comment.objects.filter(id = parent_id)
+            if parent_qs.exists() and parent_qs.count() == 1:
+                parent_obj = parent_qs.first()
 
-    return render(request, "post_item.html", { 'post': post }) 
+        new_comment, created = Comment.objects.get_or_create(
+            user = request.user,
+            content_type = content_type,
+            object_id = obj_id,
+            content = content_data,
+            parent = parent_obj
+        )
+        return redirect('/Blog/allposts/')
+
+    comments = instance.comments
+    context = {
+        'instance': instance,
+        'comments': comments,
+        'comment_form': form,        
+    }
+
+    return render(request, "post_item.html", context) 
 # delete_post
 def delete_post(request, id):
+    
     if request.method == 'POST':
         post = Post.objects.get(id = id)
-        post.delete()
+        if request.user.username == post.author:
+            post.delete()
+        else: 
+            print(123)
+            return redirect('/Blog/allposts/')   
     return redirect('/Blog/allposts/')
